Add test
Note: this test fails with polendina in a browser environment as it has to create its own http server. But it demonstrates that the code works. Once the echo/info servers used in the other tests are updated to return the correct content-type, the test can be rewritten to use those and browser tests will also pass.